Duplicate downloads

+++ yoump3.py
@@ -11,27 +11,30 @@ folderD = "C:/Users/User/Documents/Python/Download Youtube/Downloads"
 for i in range(len(lines)):
     vid = pafy.new(lines[i])
     descr = [vid.author, vid.title, vid.duration]
-    print('Autor {}, Título {}, Duraçao {}\n'.format(*descr))
+    print('Autor {}, Título {}, Duraçao {}\n'.format(*descr)) 
+    chckEx = "{}/{}.mp3".format(folderD,descr[1])
 
-    audio = vid.getbestaudio("m4a")
-    filename = audio.download(quiet=False)
+    if  os.path.isfile(chckEx) == True:
+        print("YOU ALREADY DOWNLOAD THIS FILE\n\n\n")
+    else:
+        audio = vid.getbestaudio("m4a")
+        filename = audio.download(quiet=False)
+        
+        print("\n========CONVERTING TO MP3=======")
+        m4a_audio = AudioSegment.from_file("{}.m4a".format(descr[1]), format="m4a")
+        m4a_audio.export("{}/{}.mp3".format(folderD, descr[1]), format="mp3", bitrate="128k")
 
-    
-    print("\n========CONVERTING TO MP3=======")
-    m4a_audio = AudioSegment.from_file("{}.m4a".format(descr[1]), format="m4a")
-    m4a_audio.export("{}/{}.mp3".format(folderD, descr[1]), format="mp3", bitrate="128k")
+        os.remove("{}.m4a".format(descr[1]))
 
-    os.remove("{}.m4a".format(descr[1]))
-
-    print("======SETTING THE METADATA======")
-    namArq = "{}.mp3".format(descr[1])
-    song = EasyID3("{}/{}".format(folderD, namArq))
-    descSong = namArq.replace(".mp3","")
-    artist, title = descSong.split(" - ")
-    song["title"] = title
-    song["artist"] = artist
-    song.save()
-    print("======DOWNLOAD COMPLETE======\n\n")
+        print("======SETTING THE METADATA======")
+        namArq = "{}.mp3".format(descr[1])
+        song = EasyID3("{}/{}".format(folderD, namArq))
+        descSong = namArq.replace(".mp3","")
+        artist, title = descSong.split(" - ")
+        song["title"] = title
+        song["artist"] = artist
+        song.save()
+        print("======DOWNLOAD COMPLETE======\n\n\n")
